+++ src/dda/utils/editors/interface.py
@@ -1,0 +1,39 @@
+# SPDX-FileCopyrightText: 2025-present Datadog, Inc. <dev@datadoghq.com>
+#
+# SPDX-License-Identifier: MIT
+from __future__ import annotations
+
+from abc import ABC, abstractmethod
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from dda.cli.application import Application
+
+
+class EditorInterface(ABC):
+    """
+    This interface defines the behavior of a code editor.
+    """
+
+    def __init__(self, *, app: Application, name: str) -> None:
+        self.__app = app
+        self.__name = name
+
+    @property
+    def app(self) -> Application:
+        return self.__app
+
+    @property
+    def name(self) -> str:
+        return self.__name
+
+    @abstractmethod
+    def open_via_ssh(self, *, server: str, port: int, path: str) -> None:
+        """
+        Open the editor via SSH.
+
+        Parameters:
+            server: The server to connect to.
+            port: The port to connect to.
+            path: The path to the repository to open.
+        """